Made a commandline interface to movie-app-2
Used click and rich to improve configuration of the movie-app-2 from the
command line. With this we can reload, create and delete junk previews
and thumbnails from another command line command.
  The module: movie-app-2/preload.py holds the command line application
  interface.
Also added a shortcut alias in bash_aliases for ease on using the
command line application interface `mwifi-load` is the command as per
now.