Fixed a bug and added an algorithm for caesars cypher.
The bug in mpack.bsize module: ensure passed object was a float or
atleast convertible to a float before the actual conversion in all the
conversion functions. The cypher algorithm simply rotates characters and
matching them to a specified one given some rotation length.
Just a sample example but not the actual module API. Read the module for
further understanding.

Eg: rotate(text='ABCD', length=4) => 'DEFG'