Learning Chapter 5 of Concurrency in PyAsyncio.
Today, I was cheking out a chapter in concurrency with python asyncio.
The chapter title was `Non-blocking database drivers` where I had to
setup postgres for this. I corrected some errors in the chapter code and
I am done with it.
I need alot of exposure to SQL. PostgreSQL is powerful. (OvO)