Testing AioHTTP, HTTPX and Flask
The aiohttp web framework has once again proved to be amazing, I have
managed to create an application using aiohttp seamlessly as compared to
flask which has weird naming across all its modules. AioHTTP is amazing
in terms of security and and instead of recreating the background jobs
like fastapi, it utilizes python core features along with the standard
library. It uses asyncio.create_task to create a task that functions as
a background task, amazing and simple. The simplicity of the asyncio
design makes aiohttp even easier to reason about, DRY - Do not Repeat
Yourself. The HTTPX framework has once again created its own URL library
for its own usage while there are others created only for that purpose
and do it very well like Yarl a standalone library created as part the
aio-libs libraries in which aiohttp is a part of, I will use
AioHTTP.ClientSession for this reason over HTTPX.