Testing AioHTTP Plugin acrhitecture.
The amazing AioHTTP Package works as plugin system, where the different
desired pieces are connected together to achieve the desired effect like
the aiohttp_security package which offers a security system on the
aiohttp framework while aiohttp_sessions package offers a session
management system for aiohttp. In the dataapp project stored in
live/python/dataapp I step out to test and see this plugin system by
using a simple user management with permissions system which. The
permissions have been set using IntFlags and a global data attribute
which only those with read_data permissions can read, write_data
permission can mutate and delete_data can delete the data. There are
also meta permissions which enable the holders to mutate, see and delete
users in the system. The dataplugin is in charge of storing and loading
the users and data state while aiohttp_security connects with the
dataplugin to access the permissions and gauge all users to their
respective authorised sections of the application. This application is
stateless hence ho usage of aiohttp_sessions package, it uses basicauth
for authenticating the users.