Added a few events in movie-app-2.
Added a few events for volume change, return focus to main-player on
video change, play next video onended main-player event, added a key
binding for muting the main-player.
The python `css_html_js_minify` package has a bug, sort of. When you
minify javascript eg
`
	if (age >= 18)
		console.log('You can Vote')
	else
		console.log('You cannot Vote')
`
minifies to
`
	if(age>=18);console.log('You can vote');else console.log('You
	cannot Vote')
`
which leads to a syntax error (Unexpected token else) in the browser.
A simple solution, put the one line if and else block in the same line
as the if/else keyword, ie
`
	if(age >= 18) console.log('You can Vote')
	else console.log('You cannot Vote')
`
This leads to the correct minification.
I will report this Issue on the github Repo.