Fixed Fatal Error in last LeetCode Problem.
There were a few errors in the last LeetCode solution. There was memory
allocation without deallocation and `_temp` was deallocated irregardless
of it holding a valid value and hence leaving a dungling pointer in one
of the result list node. As a solution, I created a `deleteList` function
to delete all nodes in the used list, ie the two input lists and the
solution list and for the latter bug, I wrapt it into the correct if
block when its deallocation was required.