Fuck RUST.
All this hype that rust is better is nothing short but vaporware of
lies, rust doesn't add anything to the table, the safety thing they say
is just a simple section in the compiler that checks how object go
around, this can be achieved in c++ if a static analyzer can be coded
into the compilers which I believe will soon be supported. The
ownership-borrow mechanism is akin to c++ constructors-destructors,
python __enter__-__exit__ and any language which the
compiler/interpreter promisses to do an action after some action can
achieve this, example the try{...}finally{...} in javascript can be used
to implements this contextuallike feature.
Using the RAII, c++ is memory safe, provided the programmer is carefull
of how they use the objects, c++ is even better that way.
I hereon say, FUCK YOU RUST.