Proving RUST is not necessary.
I implemented the Option in Rust in python and javascript and it is in
the std in c++ as std::optional<T>, with this I conclude that RUST is
just a rename of existing features in other languages. I will not learn
RUST no matter what, maybe if necessay in the future, but right now,
never.