Fixes here and there. BitMask error.
Fixed some few bugs involving by simple python module`mpack.flags` which
involved function`flag_off`. This was used to turn a certain flag off in
some numeric value using the XOR operator, operator`^`. The problem was
that, if the flag is off in the given numeric value, the flag would then
be turned on and if on, then off, a toggle effect that I didn't know
about.
Possible errors would include Permission elevation, State switch, etc.
Example:
	enum Perm { READ = 1 << 0, WRITE = 1 << 1, EXECUTE = 1 << 2 };
	int perm = READ & WRITE; // involves perm 1 and 2
	perm = perm ^ READ; // permission is dropped to only 2
	perm = perm ^ READ; // this activates 1
	// the intended operation was to turn off READ permission if on
To prevent this, check if the permission is on first before switching it
off, except when this toggle behaviour is intended.

Also, the module`mpack.this` cannot be imported if there is no command
line integer argument $1, fixed this by dropping the need for import
variable`sys.argv` which seemed to work perfectly in the one lined code.