Tried to publish the hi python package.
I tried to publish my beginner level python package without any success.
According to the PyPI documentation, I have to register an account to
TestPyPI and PyPI, generate API tokens then use twine to compile my
package to uploadable format, ie .tar.gz and .whl files. Use twine to
upload to the PyPI index. The problem is, this approach fails all the
time. I have tried it multiple times with very little success.
And by the way, the building process, for whatever reasons, does not
detect the yaml (pyaml) package installed in the environment and global.
I have to find out if `yaml` is part of the python standard library.
The package I'm trying to upload is `live/projects/hi`.