Fixed Bugs Today and improved on mpack
Two major errors were found on module`mpack.stream` and
module`mpack.option`. They are not exactly errors but great
improvements. For module`mpack.option`, I added a custom NONE object
that is only created once and can never be instantiated ever again,
basically, a Singleton, used it to represent an empty option. Fantastic.
Made it bullet proof by scrubing the NONETYPE class. For the
module`mpack.stream`, there was a function for getting the null_stream
that read the variable directly, not cool since the variable is null
until it is first used. The creator is some hadler for getting and
setting the null_stream which is some protected global variable as well
as others like _original_stream. Also added a swap on the StreamFiles
class.
I managed to set the emojis on my computer, though I still don't
understand how fonts work in linux, but I have a place to start,
fontconfig, some weird XML file that is like some programming language
based on how it is written.