Wrote my first test.
Never used pytest before, my first time. It was amazing thinking of the
different cases to test in the software. I wrote test for my
module`mpack.option` and the test file is `live/python/option_test.py`.
Turns out NotImplemented and NotImplementedError are two very different
objects in python. NotImplemented is a status variable while
NotImplementedError is the actual error that can be raised while the
former is meant to be returned as a signal that the function was not
called as the logic is not implemented.
The test was written using pytest and to execute it use:
$ pytest option_test.py

All tests passed, amazingly.