Learned some BASH and fixed a bug.
You guys, process substitutions and coprocesses are awesome. The way you
can asynchronously process streamed input to multiple sources for
processing and get the required results is amazing. An example of this
can be found in `live/others/bash/ch26.sh` bash script, the function
`async_proof` demonstrates this behaviour. They can also be used where
files are expected since process substitutions return files that are
essentially file descriptors in `/dev/fd/:fd` for my OS (arch linux).

What about coprocess, while process substitutions are awesome, I bet
coprocess are even more awesome. They are created using the coproc
command and they create an array named COPROC or any custom name in
which the first element is the file descriptor piped to the stdout of
the coprocess and the second element piped to the stdin on the
coprocess, the two fds can be used to communicate with the coprocesses
and amazingly, you can open multiple coprocesses for multiprocessing
tasks. Coooool.

In my utility function `ex_alias` from my last commit had a minor
missing feature, to return the args passed to it along with the new
unaliased command. This helps with less typing and also ex_alias
chaining support to be added later.