Finished asyncio chapter 7.
New tricks were learned as always. Fixed a bug in the tkinter
application example in the asyncio book, it felt awesome. The error was
pretty weird, at first I didn't know where to start but a few hops here
and there got me closer to the solution. The concept of running an
EventLoop in a different thread is mind blowing and opens way for
infinite possibilities especially with integration with synchronous
codebases.