Read on Bash PS1 customization.
Today's read was fun. I finally understand how PS1 customizations work,
how to include emojis and make it behave differently depending on the
environment. Example, if the current directory is a git repo, then PS1
should include the name of the current branch and how many commits were
made to thay repo and if not in a repo, all that info is not shown. Also
read on brace expansion and the tricks they offer. They really do make
command scripting very easy by avoiding repetition.