Read on first third of asyncio chapter 8.
The Streams topic, python asyncio chapter 8, is quite challenging as
compared to the last few chapters I've read on. The APIs, Transport
and Protocols are quite a mouthful, but they do pose a very huge
opportunity for improvement and design.
Not long ago, I implemented the `ainput` method that was akin to the
builtin `input` method, it run asynchronously by running in another
thread, but this, this Stream magic did it asynchronously in the current
event loop same thread same process without blocking. Amazing. Am yet to
understand this though, I mean this commit is full of doubt.