Read on Bash Jobs and fixed README.md
Running jobs is far much simpler than spawning new terminal instances
just to run a command in the background. The jobs API and other
associated commands like bg, fg and kill ease the process of creating
servers to run in the background of the current executing terminal
without leaving it in the first place, and don't get me started on the
idea of connecting thos jobs to sockets and named pipes to be able to
interact with them seamlessly, they remind me of services, a tool I'm
blind to and yet so powerful, jobs are to services as threads are to
processes.

Previously I had intergrated some new knoledge on markdown into my
README.md file hoping to enhance it but it turns out, github's markdown
is wildly different from vscode-markdown preview as url mappings for
relative file paths work on my machine using the caret symbol while on
github they are regarded as footnotes, this has definately trashed a big
part of my documentations. (-_-)