Finished Python asyncio chapter 8.
Finally got a deeper understanding of the asyncio streams and how to
create servers with it. Compared to handling sockets directly, streams
are much easier to deal with. Also asyncio has its own mechanism for
dealing with sockets using the loop, loop.sock_* methods are used to
create,connect,bind ... sockects then transports are built on top of
them, then protocols, then streams, then servers which are built on
streams, amazing.