Implemented the pow function in pure python.
Finally came up with the an idea on how the pow function works based on
my last discovery on binary search. I have implemented two versions
which operate on the two sides of the pow function, the whole part and
the fractional part. Instead of looking at the power as a whole
fraction, I thought of it as just rooting then powering which hugely cut
down the load, then I made a simple power function using a for loop,
which turned out to be super slow, at about O(N), I sought to improve
this, and I believe I made an algorithm that takes O(log(N)) which is
has as fast as the builtin pow, mostly because its written in c, but
that was fast enough. Then I inverted the nthroot function, and used
both implementations and wala, pow. I'm sleepy now and I feel this
commit is getting way too long even for me, why am I writing this.
$^%#@%&*