Made my pow a little bit faster.
So, I took a long walk around school thinking about the my
implementation of the pow function I did early this morning and it
occurred to me that I can improve the performance by getting rid of
the recursion and the inner for loops in those recursions by saving to
the cache the partial product of the whole product to the nth term
instead of recalculating it whenever needed, this is what the for loops
did, so instead of saving the nth terms, I save the product of the first
n terms, hence no need for recalculation and for-loops, it was a bit
tricky but I got it, the new algorithm is now faster the earlier one, it
is even faster as the exponent grows, the higher the exponent the faster
it is and it now truly works in O(log(N)), I think the earlier one was
O(N*log(N)). I don't exactly know the difference but if I had to take a
guess, due to the internal for-loops which all add up to greater than N,
then it has to be N*log(N), though I will look into it later.
HappyDays (*_*).