Done the pow(x, n) c++ leet code problem.
I have prepared for this problem for a very long time, it is the reason
for the pow.py module, my implementation of python's pow function. To
implement it in c++, I had to be very careful since issues like
overflows come from nowhere and also, in my python implementation, I
used the exponentiation operator to keep track of the exponent as it
decays but that defeats the whole reason to write my own pow function.
In c++, I came up with a new solution for this, as I stored the round's,
value, I also had to store the exponentiation at that point, did solve
the problem. I have solved this problem in my python growing functions,
removed ** for tracking the exponentiation value that was already in the
cache.