Learning how to port c++ into python.
Turns out there is alot to consider when porting some possible fast c++
code to python the first being, the types and what library to use to
port from c++ to python. There are a bunch of them out there but some of
them are just soooo hard to work with to an extent that not using them
is easier and will definately make your life easier, I am taling about
you ctypes. The ctypes std python module is one of the "easiest" ways to
bind c/c++ shared libraries into python but it seems to be selective, it
works on some machines and not in others, ie, my first attempt of using
ctypes was on ubuntu and it worked fine but now, the same exact approach
won't work on arch, I don't know why though, the second and best from my
POV is using boost_python, except for new compiler flags I had to learn,
it was easier to understand what was going on compared to ctypes,
pybind11 and using python.h directly.
I ported my earlier c++ pow solution for python, the output result, for
the most part was veeery inaccurate due to narrowing from c++ long to
python int which is immensely big, other than that, c++ was pretty fast,
twice as fast as the python builtin pow, later I will try to consider
the types for the c++ solution, since boost provides python bindings types
like python int equivalent to boost::python::long_, I
don't quite understand them right now.