Did 1 leetcode problem and tested immutability
Todays leet code problem was hard, but ranked in the medium tier. It
involved creating a fast food ranking system, it seemed easy at first
but using wrong approach does lead to some expensive operations such as
looping the whole array of foods, ratings and cuisines just to find what
you are looking for, this can be simplified using hash tables, a little
like indexing in database such that, organising the data into the hash
table is expensive at first but searching and iterating the structure
becomes easier, not easy. The problem is, even after using hash tables,
with a syfficiently large dataset, the original problem comes back to
life, example, consider 1000 foods and their ratings in the kenyan
cuisine mapping along with others like ugandan, tanzanian and ethiopian.
To find the highest rated food in those lists becomes O(N) and this
cannot be optimized away by sorting since there should be an option to
change the rating of a food in a certain cuisine hence the sorting
behaviour should be dynamic in which redoing also makes everything super
slow, and this is where the good ol' heaps shine. I have a solution in
mind, not yet implemented but will be asap.

The idea of creating immutable classes is amazing, I implemented one
today with the use of the __new__ method as defined by classes, amazing.
Check it out.