Read on asyncio chapter 12 and made fisrt Issue.
Today I read on asynchronous Queues and how they can be used in producer
consumer workflows which was very facinating, there were background
workers that dequeued objects/tasks and processed them while
clients/producers enqueued objects/tasks for processing, this sounds
like event processing.

Also read on the Barrier synchronization primitive, it uses the dreaded
Condition, but it is easier to comprehend, as it blocks tasks until they
reach a certain magnitude then simultaneously unblocks them to continue,
this reminds me of the games that are started when there are enough
players to play the game, I guess this is one use case for this.

My first issue, involved aiohttp, very unclear as it may also be a
python issue, I don't really know what exactly is the problem but sooner
or later, I will figure it out.