Completed the Issue and fixed webrun errors.
The aiohttp issue was completed and it turns out, it was my computers
configuration all along, I still haven't figured out what the problem is
exactly but it is definately not python, asyncio or aiohtp as confirmed
by Sam.

The webrun module has been throwing all types of bugs mainly due to poor
implementation, the main issue was trying to parse the string input
which proved very hard and errorneous. I replaced that part with regex
and captured necessary errors along the way to its completion. The new
implementation works fine, I will test it later though.