Solved one leetcode problem and created `gitrepo`
Todays leetcode was pretty tricky and damn hard, at first. It needed a
complete rethink of how I view loops, not that much, but a big part of
the solution depended on that. I managed to design a solution that works
on any valid input, the predefined limits are a subset of what it can
do. Sooo good that I found an issue with one of the testcases and made a
github issue on the leetcode forum account. 2nd issue now.

As for the `gitrepo`, just a simple command to get the repository in
which a path is part of, default directory if not passed.