Simple logging module and better naming.
Today's work is centered around naming and implementing a simple logging
library compared to the default logging for python. It has some simple
interface that's easy to use and understand with some probably unnecessary
features like logging states, which simple say, what was on when a log
was written, like all things marked on when the log was written.

The better namings part is due to the poor naming in mpack.flags module
where I replaced flag_on, flag_off, flag_enabled and flag_disabled with
turnon, turnoff, ison, and isoff respectively. I find the new names better
and easily convey what the functions do in addition to the new docstrings
I've added.