Finally, finished python asyncio book.
The asyncio book, complete, done, kapish. Learned a lot about asyncio,
met some really cool programming techniques and found a few reasons to
appreciate linux over windows, eg. unix-sockets and others, they are
brilliant ideas. Still have a lot to learn about asyncio, it is one
hell of a library, "what is await asyncio.sleep(0)" I first thought,
oooooh, force an event loop iteration, it is like a way of telling the
event loop to push the current task to the end of the task list and
continue running the tasks till it gets back to the task, as usual this
is not all, its just the simple understanding I have, I will create a
simple event loop made purely of generators, nothing fancy, actually, I
had already made one, but it was veeery poorly designed. With my new
understanding of asyncio tricks and tracks, here and there, I should be
able to spin one up, that's easier on the eyes and mind. Booooyaaa!