A little event loop on python generators.
I have implemented a toy event loop based on python generators, it is
simply a proof of concept to see if generators have what it takes to
create an event loop, it is amazing what you can do with this stuff. The
generators seem to handle everything very well, a more robust error
handling is one of the main requirements of the one file project.