Read on State Design Pattern.
I remember the day I tried to create a simple GUI application with
multiple pages from Login, to the main page with a dashboard that had
buttons that could change what was in view in the main content area, it
was hard and pretttttty difficult. I had no way of controlling which
object was in view and which one was not, hundreds of possible state all
intertwined together and most of the properties were invalid from state
to state due to improper cleanup when transitioning from one state to
the other. But with this new design pattern, I can clearly see the
application working from login to shutdown with different procedures
running when a state comes to life.
I have a lot to learn on this design, it seems pretty elegant. I hear it
uses Finite State Machines, whatever that is, I'll read on it later.