Read on KNN algorithm.
Well, KNN, simple beautiful and interesting algorithm. You can do some
interesting things with a well defined dataset. I have implemented a
simple KNN algorithm that classifies fruits -- oranges and grapes. The
two can be defined numerically by two integers, color and size in the
range of 1 to 10000. where the color is from orange (1) to purple(10000)
and the size from small (1) to large(10000). I have generated some
simple random dataset that can be used to predict a fruit as either an
orange or grape given its characteristics, size and color. big and
orange, probably and orange, small and purple, probably a grape. This is
just a simple illustration of the data. You can play with it and plot it
again and again and also, generate new data. Pass in some random points
and see what type of fruit they mostly represent. Control k and do
whatever you want with it, see the effects of a small k, big k, blah
blah blah.