Basic C++ iterator protocol.
At first I thought the iterator protocol in c++ was only confined to
contiguous memory blocks, there begin returns the start memory location
and end, the address after the last memory location. Then it hit me, c++
supports operator overloading, and to confirm, there are about 3
operations done on the memory being iterated on, one, check if we are at the
end, two (increment) move one step forward, and three dereference the
current address to get value in that location which is the target of the
iteration in the first place. These three are achieved using the
operators `addr == endaddr`, `++addr` and `*addr`.

If I was to create an iterable that returns an object storing a state
and a special stop token with the target operators overloaded then I
would be able to iterate the iterable. For this experiment, I tried to
implement the python range object in c++. It simply returns range state
aptly named `rstate`. where, `range.begin` returns an `rstate` and `end`
which creates a special `rstate` which simply marks the end of the
iterable created by `range.begin`. I had a lot of fun implementing this.
To think this simple interface would mean so much if you thought about
it.

I've heard of `std::iterator_traits` they seem complex. I will later
read on it, for now, am satisfied with my little understanding. And that
is not even all, generators are coming. DOPE!!!