Background Work. Read on math, c++ templates, etc.
Once upon a time, I tried to read later chapters of the `C++ template
metaprogramming` by David, specifically the topic on `typelists` and
operations on them, I failed miserably. But the content kept running
around my brain since them and yesterday, it snapped, `that is what is
was`. Out of nowhere I understood the contents of the type list and made
an effort to write my own from scratch, not bragging or anything, it
works perfectly. The most intense operation to implement was the
reverse, to reverse the type list. Partial specializations were also a
mistery that uncovered yesterday.
Otherwise, today I did some linear algebra, played around with lazy
evaluatins, which are pretty cool, they are used in sympy library. Can't
really say how, but I made a simple one, thanks to crafting interpreters
introduction to tree walk interpreteros that really helped in the
understanding of lazy evaluations.
I've always wondered how python knows when an operator is not defined
for, example __add__, where it will then try to call, for this case,
__radd__. I've experimented on a few things here and there and I've come
to a conclusion, if the operator, example __add__, returns the
NotImplemented object then __radd__ is tried, if it also returns the
NotImplemented object, an error, TypeError, is raised that the +
operation is not defined for the operands provided. Cool. This just blew
a whole lot of ideas in my head.