A little more of click-ing, you see that (*_*)
Squeezed in todays work by reading on passing passwords in click and the
prompt kwarg together with the confirmation_prompt. You can also use the
password_option decorator since asking for passwords is common, it was
created. This made me think of plugins, does click support plugins, I'm
not sure but it would be cool if that was possible, I'll look into this.