More math for you.
Old days, PI would be calculated by dividing inscribing a polygon with n
sides into inside a circle, calculate the area of the polygon by area of
each triangle times the number of sides to get an approximate area of
the circle, the bigger n was, the more accurate the result would be. Now
with this approximate area of the circle, divide it by radius squared
and you would end up with PI, (ussually the radius was 1 and therefore
the area would be PI). This was painstakingly low and tedius, some guy
computed the first 35 digits of PI for the major part of his life,
yikes. The Newton cameby and fucked them all up by creating calculus,
integrating the function of a circle `x^2 + y^2 = r^2`, and computing
the integral from x=0 to x=r, then multiply by 4 to get the area of the
circle and divide by r^2 to get PI, (again, r is 1 for most of this
calculation, for simplification) the PI formula was now

                      / r                 \
	              | /   __________    |
	          4 * | | \/ r^2 - x^2 dx |
	              | /                 |
	              \ 0                 /
	     PI = -----------------------------
	                       r^2

With this, you can calculate PI to the 100th digit in a matter of weeks
rather than a whole lifetime, I feel so sorry for the poor guy.
Today, I computed PI using calculus, linear algebra and the old polygon
style and one thing is, the polygon style is pretty tedius. Infact I
cannot pass the 5th digit of PI in accuracy, the rest give me an
accurate answer as much as math.pi is concerned. Now, in my program, I
have used sin, cos which operate on radians meaning I had to interact
with math.pi and math.tau somewhere but assume they have been conputed
using tailor series, with degrees or something. This was todays work,
mostly math on Paper.