A tiny dose of javascript and admiration for alpine
Javascript is a weird language and yet amazing, the prototype thingy
going on for it is very questionable, I've tried to tap on this
prototype thingy and I've failed in all attempts, but I have concluded
one thing, everything, from user defined objects to constructing new
objects using the new and the class objects, they're all just
dictionaries (mappings from some name to some value), I know, not very
amusing but it makes alot of sense now. Example:

	function Person(name, age) {
		this.name = name
		this.age = age
		this.greet = () => console.log(`Hey ${this.name}`)
	}
	let me = new Person('My Name', 37)
	// is similar to doing
	let me2 = {}
	Person.call(me2, 'My Other Name', 37)
	assert.equal(me2.age, me.age)

This is not much, but it does illustrate the objects and everything
being some object with properties, the new keyword is simply a syntactic
sugar for binding some object to this and calling Person then that
object, now filled with the properties is returned, I'm yet to
understand the weirdness of javascript but I think this is a pretty good
start. Though the classes have a benefit, unlike the function style,
classes preserve the methods and call them by binding the instance then
the call while the functions create new function objects which are
stored are properties on the object, I've not got that right probably
but its what it is, visit the oop.js module in the javasript folder to
see what I mean. Long Day, whew!