More exploration on alpine, htmx and axios
I don't know if it is my pure hate for javascript or are these libraries
just absurd. No type hints whatsoever, nothing! I have to install
extensions to highlight them especially when it is a script tag in html.
Though, I have learned some new tricks in javascript,
`Object.defineProperty` is a way to set getters and setters in the good
old function objects.
I do not understand why the alpine plugins have been made in a special
way, they should just be plugins like others, I don't like that at all.
I'm not sure with alpine as the developer seems to be maddened with
adding features to it, its good for growth but if this keeps going on,
it will morph into react/vue/angular and others.
I have a suggestion, make alpine just act on the core data and its
behaviour then the plugins should carry all the other unnecessary
features and addons, the laravel thing they are tring to add should not
be in there, should be a plugin that laravel users should install to get
an interface with alpine from laravel framework, what am I saying, I do
not know how to program massive projects but I have a taste for these
things, ignore my thoughts.