Type punning and some other things here and there
I promissed to look at type punning in c from the languages project and
I have fulfilled that by testing its capabilities and am I impressed.
Though not thoroughly tested, I have managed to skim over the
capabilities of the feature and it does seem to unlock and explain alot
on how large systems are built using c. As a test for the type of power
am going for, I will try to write a simple c json parser. This feature
allows one to pass around abitrary types of data. For example, if the OS
was to send an event to your application where there are 5 possible
types of events, the OS allocates the memory and give it to me to free,
I can untag the object to get the actual type of event then react
appropriately. You can also have a list of arbitratry types of data
under a common type of object with a tag that represents the underlying
value it holds.
Otherwise, learned some new way of constructing objects in C, it almost
looks like object oriented.