Pointer arithmetic, packing and alignment of memory
I have always wondered, if I had the address of a private field in some
class for some instance, does setting a value to the deferenced addrsss
outside the class possible? Yes, actually, a class/struct in c/c++ is
simply a blob of data where the fields are stored one after the other as
they were defined, therefore the block after the block of memory of
the first field is the block of the second field, at least if the
struct/class has been aligned/packed. To align a struct in memory you
can use `__attribute__((packed))` after the struct/class keyword, this
also requires that the types of the fields are also aligned to get a
perfect alignment, why is it not aligned by default? My theory is for
types like int, which is 4 bytes on my machine, are padded to fill a
whole block, 8 bytes I assume, so that the int has an address equal to a
block, I think this is some hardware optimization or OS. I don't know. I
have documented a whole lot in the `*align*.cpp` files, readon in there
to get the whole picture. I've also checked out a way of overloading
functions by tagging them, by using a set of possible tags to represent
different algorithms of doing something, you can tag each algorithm such
that when called, a specific version is picked depending on the tag it
holds or explicitly passed hence overload. I do not have a clear
implementation but it is good enough for a basic picture, usually tag
dispatch uses std::enable_if with the tag being a type or something.