Tested the extern C thingy and fixed some bugs
I was thinking of extern "C" as used when including C header files in
C++. One thought came to mind, did it also supress the use of C++
keywords in the C source, like, did it strip the meaning of the wrongly
used C++ keywords so that the C code was valid C++ code. If in the
C headers, 'this' was used as a variable in some function and 'class' as
a member in some struct. Turns out extern C does not even think of them.
Using a little magic of preprocessing, we can convert the invalid code
to some correct C++. By replacing this to this_ and class to class_ and
wrapping it in extern C and a namespace to prevent global polution, we
have successfully included C header file in C++ while maitaining its
overall theme. But I still don't know what the extern C is all about. As
for the bugs, It was about a subtle difference in object contruction
from C to C++. In C++ you can convert an initializer list to the
intended aggregate type but for C, this has to be explicit, like a cast
from some list to the type which I forgot in the vector.h file. BYE.