Learned some new C trick, awesome.
This commit is all about c/c++ operators, not the advanced kind though,
just the good ol' + and -. Heres the thing, doing a pre
increment/decrement operator on a number variable returns the new value
in the number while the post version of the operators on the number
variable return the value before the operation. Example:
	int number = 100;
	std::cout << number++; // 100  number=101
	std::cout << ++number; // 102  number=102
This was a bug in the languages c_lox project, I didn't know of this and
it was hard to track I tell you. I had to read the source like I was a
beginner to C, just to pin it down, but now that I know, I'm gonna use
it. This is not all, turn out that C++ kind of doesn't have
commutativity for user defined operator overloads, obvious but I thought
for externally defined overloads would be commutative, but it's good to
know. I downloaded the python dev source code, like the C source and I
have to say, they use goto alot. But if you think about it, goto is kind
of a jump instruction and thats all it is, therefore not a bad feature.
The goodness is in how it is used, poor usage leads to unreadable code.
Thats all.